Issue #1065: Lazy-load apiKey/proxySigned authenticators
